+++ src/transformation/utils/safe-names.ts
@@ -42,7 +42,6 @@ const luaBuiltins: ReadonlySet<string> = new Set([
     "pcall",
     "print",
     "rawget",
-    "rawset",
     "repeat",
     "require",
     "self",