+++ src/transformation/visitors/class/utils.ts
@@ -6,20 +6,21 @@ export function isStaticNode(node: ts.Node): boolean {
     return (node.modifiers ?? []).some(m => m.kind === ts.SyntaxKind.StaticKeyword);
 }
 
-export function getExtendedTypeNode(
+export function getExtendsClause(node: ts.ClassLikeDeclarationBase): ts.HeritageClause | undefined {
+    return (node.heritageClauses ?? []).find(clause => clause.token === ts.SyntaxKind.ExtendsKeyword);
+}
+
+export function getExtendedNode(
     context: TransformationContext,
     node: ts.ClassLikeDeclarationBase
 ): ts.ExpressionWithTypeArguments | undefined {
-    if (node && node.heritageClauses) {
-        for (const clause of node.heritageClauses) {
-            if (clause.token === ts.SyntaxKind.ExtendsKeyword) {
-                const superType = context.checker.getTypeAtLocation(clause.types[0]);
-                const annotations = getTypeAnnotations(superType);
-                if (!annotations.has(AnnotationKind.PureAbstract)) {
-                    return clause.types[0];
-                }
-            }
-        }
+    const extendsClause = getExtendsClause(node);
+    if (!extendsClause) return;
+
+    const superType = context.checker.getTypeAtLocation(extendsClause.types[0]);
+    const annotations = getTypeAnnotations(superType);
+    if (!annotations.has(AnnotationKind.PureAbstract)) {
+        return extendsClause.types[0];
     }
 }
 
@@ -27,6 +28,6 @@ export function getExtendedType(
     context: TransformationContext,
     node: ts.ClassLikeDeclarationBase
 ): ts.Type | undefined {
-    const extendedTypeNode = getExtendedTypeNode(context, node);
-    return extendedTypeNode && context.checker.getTypeAtLocation(extendedTypeNode);
+    const extendedNode = getExtendedNode(context, node);
+    return extendedNode && context.checker.getTypeAtLocation(extendedNode);
 }